requirements.txt
Flask==1.1.2
Jinja2==2.11.2
Pillow==7.1.1
